python-backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.131.0",
    "openai-agents>=0.9.3",
    "openai-chatkit>=1.6.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
